Extract search into separate source
